requirements.txt
selenium
requests
python-dotenv
beautifulsoup4
tqdm
pytest-xdist
browserstack-local